requirements.txt
appdirs==1.4.4
CacheControl==0.12.6
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
distlib==0.3.1
distro==1.5.0
Flask==1.1.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==1.0.0
packaging==20.7
pep517==0.9.1
progress==1.5
PyMySQL==0.10.1
pyparsing==2.4.7
python-dotenv==0.15.0
pytoml==0.1.21
requests==2.25.0
retrying==1.3.3
six==1.15.0
toml==0.10.2
urllib3==1.26.2
webencodings==0.5.1
Werkzeug==1.0.1
